requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain-groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette